Added support for searching by administration area or country
The input string "name" is split by comma character (","), first part is used for the original search, seconds part is used for separate search for administration areas and countries.
This can be used for example to search for city Queenstown in New Zealand with query "/v1/search?name=Queenstown,New Zealand".
If the input string doesn't contain comma, it works as previously.